Keep ert-tests/ of the load-path
A file named ".nosearch" keeps the containing directory of the
`load-path', which is necessary here because it contains a bundled
library ("ert.el").

Bundling third-party libraries is problematic because some people
install packages directly from their Git repository and that could
lead to the bundled (and potentially outdated) copies of the bundled
library shadowing the upstream version.